refactor(bundler): align .config webpack helpers with scaffold layout
Adopt the create-plugin 7.9.0 scaffold's bundler structure:
- move .config/webpack/{constants,utils}.ts -> .config/bundler/
- add .config/bundler/copyFiles.ts and use its copyFilePatterns in
  webpack.config.ts instead of the inline CopyWebpackPlugin patterns
- drop the now-unused hasReadme import from webpack.config.ts

The new copyFiles.ts copies only plugin.json-declared logos/screenshots
(plus README/LICENSE/CHANGELOG/plugin.json/**.json/query_help.md) rather
than the previous broad **/*.svg|png|html + img/libs/static globs.

Verified byte-for-byte: built dist/ before and after the refactor and
diffed manifests (sha256 + size per file) -- identical file set, no
assets added or dropped. src/img/{logo.svg,screenshot.png} are both
declared in plugin.json so they remain covered. typecheck, lint, and
unit tests (170) all pass.
